@@ -1,8 +1,14 @@
 module github.com/piprate/json-gold
 
-go 1.13
+go 1.18
 
 require (
 	github.com/pquerna/cachecontrol v0.0.0-20180517163645-1555304b9b35
-	github.com/stretchr/testify v1.4.0
+	github.com/stretchr/testify v1.8.3
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
