@@ -19,10 +19,10 @@ package ld
 // underlying algorithms
 //
 // Warning: using this interface directly is highly discouraged. Please use JsonLdProcessor instead.
-type JsonLdApi struct {
+type JsonLdApi struct { //nolint:stylecheck
 }
 
 // NewJsonLdApi creates a new instance of JsonLdApi.
-func NewJsonLdApi() *JsonLdApi {
+func NewJsonLdApi() *JsonLdApi { //nolint:stylecheck
 	return &JsonLdApi{}
 }
