@@ -15,9 +15,10 @@
 package ld_test
 
 import (
+	"testing"
+
 	. "github.com/piprate/json-gold/ld"
 	"github.com/stretchr/testify/assert"
-	"testing"
 )
 
 func TestGetFrameFlag(t *testing.T) {
